Markdown support
Implement Markdown support with the Blackfriday. This support is both
for natively rendering Markdown files, as well as exposing markdown
rendering to Janet.

Also implemented Go templating in both HTML and Markdown. I may opt to
remove this, or replace it with Markdown plus Temple somehow.
