+++ examples/hello.md
@@ -1,0 +1,4 @@
+# Hello from Markdown
+
+Spinnerette also supports Markdown rendering!
+
