+++ go.mod
@@ -1,3 +1,5 @@
 module github.com/rushsteve1/spinnerette
 
 go 1.17
+
+require github.com/russross/blackfriday/v2 v2.1.0
